refactor and remove cruft
