+++ src/Particular.CodeRules.Tests/Helpers/AnalyzerTestFixture.cs
@@ -10,67 +10,55 @@
     using Microsoft.CodeAnalysis.Text;
     using Xunit;
 
-    public class CSharpAnalyzerTestFixture<TAnalyzer> : AnalyzerTestFixture where TAnalyzer : DiagnosticAnalyzer, new()
+    public abstract class AnalyzerTestFixture<TAnalyzer> : TestFixture where TAnalyzer : DiagnosticAnalyzer, new()
     {
-        protected override string LanguageName => LanguageNames.CSharp;
-
-        protected override DiagnosticAnalyzer CreateAnalyzer() => new TAnalyzer();
-    }
-
-    public abstract class AnalyzerTestFixture : BaseTestFixture
-    {
-        protected abstract DiagnosticAnalyzer CreateAnalyzer();
-
-        protected Task NoDiagnostic(string code, string diagnosticId)
+        protected async Task NoDiagnostic(string code, string diagnosticId)
         {
             var document = TestHelpers.GetDocument(code, LanguageName);
-            return NoDiagnostic(document, diagnosticId);
-        }
-
-        protected async Task NoDiagnostic(Document document, string diagnosticId)
-        {
             var diagnostics = await GetDiagnostics(document);
+
             Assert.Empty(diagnostics);
         }
 
-        protected Task HasDiagnostic(string markupCode, string diagnosticId)
+        protected async Task HasDiagnostic(string markupCode, string diagnosticId)
         {
             Assert.True(TestHelpers.TryGetDocumentAndSpanFromMarkup(markupCode, LanguageName, out Document document, out TextSpan span), "No markup detected in test code.");
 
-            return HasDiagnostic(document, span, diagnosticId);
-        }
-
-        protected async Task HasDiagnostic(Document document, TextSpan span, string diagnosticId)
-        {
             var diagnostics = await GetDiagnostics(document);
-            Assert.Single(diagnostics);
+
+            _ = Assert.Single(diagnostics);
 
             var diagnostic = diagnostics[0];
+
             Assert.Equal(diagnosticId, diagnostic.Id);
             Assert.True(diagnostic.Location.IsInSource, "Diagnostic not in test code.");
             Assert.Equal(span, diagnostic.Location.SourceSpan);
         }
 
-        async Task<ImmutableArray<Diagnostic>> GetDiagnostics(Document document)
+        static async Task<ImmutableArray<Diagnostic>> GetDiagnostics(Document document)
         {
-            var analyzers = ImmutableArray.Create(CreateAnalyzer());
+            var analyzers = ImmutableArray.Create<DiagnosticAnalyzer>(new TAnalyzer());
             var exceptions = new List<ExceptionDispatchInfo>();
             var compilation = await document.Project.GetCompilationAsync(CancellationToken.None);
+
             var analyzerOptions = new CompilationWithAnalyzersOptions(
                 new AnalyzerOptions(ImmutableArray<AdditionalText>.Empty),
                 (exception, analyzer, diagnostic) => exceptions.Add(ExceptionDispatchInfo.Capture(exception)),
                 false,
                 false,
                 false);
+
             var compilationWithAnalyzers = new CompilationWithAnalyzers(compilation, analyzers, analyzerOptions);
             var discarded = compilation.GetDiagnostics(CancellationToken.None);
 
             var tree = await document.GetSyntaxTreeAsync(CancellationToken.None);
 
             var builder = ImmutableArray.CreateBuilder<Diagnostic>();
+
             foreach (var diagnostic in await compilationWithAnalyzers.GetAnalyzerDiagnosticsAsync())
             {
                 var location = diagnostic.Location;
+
                 if (location.IsInSource && location.SourceTree == tree)
                 {
                     builder.Add(diagnostic);
