+++ src/Particular.CodeRules.Tests/Helpers/TestFixture.cs
@@ -1,6 +1,6 @@
 ﻿namespace Particular.CodeRules.Tests
 {
-    public abstract class BaseTestFixture
+    public abstract class TestFixture
     {
         protected abstract string LanguageName { get; }
     }
