+++ src/Particular.CodeRules.Tests/Helpers/TestHelpers.cs
@@ -10,7 +10,7 @@
 
     public static class TestHelpers
     {
-        public static bool TryGetCodeAndSpanFromMarkup(string markupCode, out string code, out TextSpan span)
+        static bool TryGetCodeAndSpanFromMarkup(string markupCode, out string code, out TextSpan span)
         {
             code = null;
             span = default;
@@ -23,7 +23,7 @@
                 return false;
             }
 
-            builder.Append(markupCode.Substring(0, start));
+            _ = builder.Append(markupCode.Substring(0, start));
 
             var end = markupCode.IndexOf("|]");
             if (end < 0)
@@ -31,8 +31,9 @@
                 return false;
             }
 
-            builder.Append(markupCode.Substring(start + 2, end - start - 2));
-            builder.Append(markupCode.Substring(end + 2));
+            _ = builder
+                .Append(markupCode.Substring(start + 2, end - start - 2));
+            _ = builder.Append(markupCode.Substring(end + 2));
 
             code = builder.ToString();
             span = TextSpan.FromBounds(start, end - 2);
@@ -42,30 +43,21 @@
 
         public static bool TryGetDocumentAndSpanFromMarkup(string markupCode, string languageName, out Document document, out TextSpan span)
         {
-            return TryGetDocumentAndSpanFromMarkup(markupCode, languageName, null, out document, out span);
-        }
-
-        public static bool TryGetDocumentAndSpanFromMarkup(string markupCode, string languageName, ImmutableList<MetadataReference> references, out Document document, out TextSpan span)
-        {
             if (!TryGetCodeAndSpanFromMarkup(markupCode, out string code, out span))
             {
                 document = null;
                 return false;
             }
 
-            document = GetDocument(code, languageName, references);
+            document = GetDocument(code, languageName);
             return true;
         }
 
-        public static Document GetDocument(string code, string languageName, ImmutableList<MetadataReference> references = null)
+        public static Document GetDocument(string code, string languageName)
         {
-#pragma warning disable IDE0079 // Remove unnecessary suppression
-#pragma warning disable IDE0054 // False positive
-            references = references ?? ImmutableList.Create<MetadataReference>(
-                MetadataReference.CreateFromFile(typeof(object).GetTypeInfo().Assembly.GetLocation()),
-                MetadataReference.CreateFromFile(typeof(Enumerable).GetTypeInfo().Assembly.GetLocation()));
-#pragma warning restore IDE0054 // False positive
-#pragma warning restore IDE0079 // Remove unnecessary suppression
+            var references = ImmutableList.Create<MetadataReference>(
+                MetadataReference.CreateFromFile(typeof(object).GetTypeInfo().Assembly.Location),
+                MetadataReference.CreateFromFile(typeof(Enumerable).GetTypeInfo().Assembly.Location));
 
             return new AdhocWorkspace()
                 .AddProject("TestProject", languageName)
