+++ src/Particular.CodeRules/Particular.CodeRules.csproj
@@ -15,7 +15,6 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Workspaces" Version="3.9.0" PrivateAssets="All" />
     <PackageReference Include="Particular.Packaging" Version="1.1.0" PrivateAssets="All" />
-    <PackageReference Update="NETStandard.Library" PrivateAssets="All" />
   </ItemGroup>
 
   <ItemGroup>
@@ -27,7 +26,6 @@
 
   <ItemGroup>
     <None Include="$(OutputPath)\$(AssemblyName).dll" Pack="true" PackagePath="analyzers/dotnet/cs" Link="$(AssemblyName).dll" Visible="false" />
-    <None Update="tools\*.ps1" Pack="true" PackagePath="tools\%(Filename)%(Extension)" />
   </ItemGroup>
 
 </Project>
