--- src/Particular.CodeRules/tools/uninstall.ps1
@@ -1,65 +1,0 @@
-param($installPath, $toolsPath, $package, $project)
-
-if($project.Object.SupportsPackageDependencyResolution)
-{
-    if($project.Object.SupportsPackageDependencyResolution())
-    {
-        # Do not uninstall analyzers via uninstall.ps1, instead let the project system handle it.
-        return
-    }
-}
-
-$analyzersPaths = Join-Path (Join-Path (Split-Path -Path $toolsPath -Parent) "analyzers") * -Resolve
-
-foreach($analyzersPath in $analyzersPaths)
-{
-    # Uninstall the language agnostic analyzers.
-    if (Test-Path $analyzersPath)
-    {
-        foreach ($analyzerFilePath in Get-ChildItem -Path "$analyzersPath\*.dll" -Exclude *.resources.dll)
-        {
-            if($project.Object.AnalyzerReferences)
-            {
-                $project.Object.AnalyzerReferences.Remove($analyzerFilePath.FullName)
-            }
-        }
-    }
-}
-
-# $project.Type gives the language name like (C# or VB.NET)
-$languageFolder = ""
-if($project.Type -eq "C#")
-{
-    $languageFolder = "cs"
-}
-if($project.Type -eq "VB.NET")
-{
-    $languageFolder = "vb"
-}
-if($languageFolder -eq "")
-{
-    return
-}
-
-foreach($analyzersPath in $analyzersPaths)
-{
-    # Uninstall language specific analyzers.
-    $languageAnalyzersPath = join-path $analyzersPath $languageFolder
-    if (Test-Path $languageAnalyzersPath)
-    {
-        foreach ($analyzerFilePath in Get-ChildItem -Path "$languageAnalyzersPath\*.dll" -Exclude *.resources.dll)
-        {
-            if($project.Object.AnalyzerReferences)
-            {
-                try
-                {
-                    $project.Object.AnalyzerReferences.Remove($analyzerFilePath.FullName)
-                }
-                catch
-                {
-
-                }
-            }
-        }
-    }
-}
